Added configuration setting and autostart entry to start the tray app automatically
The setting is handled by KConfigSkeleton, which may also be used for any
settings specific to the tray app, that might be added at some future point.
Consequently a big part of this changeset is the port of the configuration
dialog from KDialog to KConfigDialog to handle the KConfigSkeleton settings
properly in the user interface.  An added bonus is, that the defaults button is
now automatically disabled, if the user interface shows the default settings.