Add a view class for MouseDevicesModel
Moving to a Qt property with a notify signal, and a separate view class are
required to use the mouse devices model together with ConfigurationWidgetMixin.
QStringList is used, because Qt properties with a standard python list fail on
roundtrip with "QObject.property" and "QObject.setProperty".  PyQt uses a user
type variant for Qt properties with python types, but a list is converted to a
variant list variant, which makes property assignment through
"QObject.setProperty()" fail.