Fixed touchpad state management
The state machine uses more granular states now to manage the touchpad.
Manually switching off the touchpad now overrides everything else.  Moreover
the mouse manager is started upon entry of the initial state, so that it will
immediately transition to "automatically off" state, if a mouse is plugged in
before synaptiks is started.